[MOV] product_expiry: reorganizing the directory structure
