+++ addons/product_expiry/__openerp__.py
@@ -4,7 +4,7 @@
     'version' : '1.0',
     'category' : 'Specific Industry Applications',
     'depends' : ['stock'],
-    'demo' : ['product_expiry_demo.xml'],
+    'demo' : ['data/product_expiry_demo.xml'],
     'description': """
 Track different dates on products and production lots.
 ======================================================
@@ -18,7 +18,8 @@ Following dates can be tracked:
 
 Also implements the removal strategy First Expiry First Out (FEFO) widely used, for example, in food industries.
 """,
-    'data' : ['product_expiry_view.xml', 'product_expiry_data.xml'],
+    'data': ['views/product_expiry_view.xml',
+             'data/product_expiry_data.xml'],
     'auto_install': False,
     'installable': True,
 }
