+++ addons/product_expiry/views/product_expiry_view.xml
